mypy/ruff linting
