+++ Makefile
@@ -16,6 +16,7 @@ install:
 	python3 -m pip install .
 
 lint:
+	$(POETRY) run ruff check . 
 	$(POETRY) run mypy --ignore-missing-imports --install-types --non-interactive plotdigitizer tests
 
 fmt:
