+++ plotdigitizer/__init__.py
@@ -1,0 +1,5 @@
+import logging
+import os
+
+LOGLEVEL = os.environ.get("LOGLEVEL", "INFO").upper()
+logging.basicConfig(level=LOGLEVEL)
