+++ tests/test_cli.py
@@ -1,12 +1,10 @@
 __author__ = "Dilawar Singh"
 __email__ = "dilawar@subcom.tech"
 
-import sys
 import subprocess
 import numpy as np
 from pathlib import Path
 
-import plotdigitizer
 
 HERE = Path(__file__).parent.resolve()
